wgengine/magicsock: factor out more separable parts
Updates #8720
