+++ wgengine/magicsock/debugknobs_stubs.go
@@ -11,6 +11,7 @@ import "tailscale.com/types/opt"
 //
 // They're inlinable and the linker can deadcode that's guarded by them to make
 // smaller binaries.
+func debugBindSocket() bool            { return false }
 func debugDisco() bool                 { return false }
 func debugOmitLocalAddresses() bool    { return false }
 func logDerpVerbose() bool             { return false }
@@ -22,4 +23,5 @@ func debugSendCallMeUnknownPeer() bool { return false }
 func debugUseDERPAddr() string         { return "" }
 func debugUseDerpRouteEnv() string     { return "" }
 func debugUseDerpRoute() opt.Bool      { return "" }
+func debugRingBufferMaxSizeBytes() int { return 0 }
 func inTest() bool                     { return false }
